requirements.txt
pandas==0.24.2
ipywidgets==7.5.0
gmaps==0.8.4
geopy==1.20.0
selenium==3.141.0
numpy==1.16.2
beautifulsoup4==4.7.1
PyYAML==5.1.1
